🚧 MY3BW9 task: split direct closeout regression coverage
